Reorganizing from test to 'newtc' tool.

+++ Frames/Objects.h
@@ -351,8 +351,8 @@ extern	ArrayIndex	Length(Ref obj);		// Length in bytes or slots
 // MapSlots calls a function on each slot of an array or frame object, giving it
 // the tag (integer or symbol) and contents of each slot.  "Anything" is passed to
 // func.  If func returns anything but NILREF, MapSlots terminates.
-typedef	Ref	 (*MapSlotsFunction)(RefArg tag, RefArg value, unsigned long anything);
-extern	void		MapSlots(RefArg obj, MapSlotsFunction func, unsigned long anything);
+typedef	Ref	 (*MapSlotsFunction)(RefArg tag, RefArg value, uintptr_t anything);
+extern	void		MapSlots(RefArg obj, MapSlotsFunction func, uintptr_t anything);
 extern	void		RemoveSlot(RefArg frame, RefArg tag);
 extern	void		SetFramePath(RefArg obj, RefArg thePath, RefArg value);
 extern	void		SetFrameSlot(RefArg obj, RefArg slot, RefArg value);
